+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [client] FeatureFlagging Get with user authorization
+
 ### Changed
 
 - Update go 1.17 -> 1.18